Consolidate style-related code into `flyout_shared.styles.ts`
